+++ src/util/common.js
@@ -68,3 +68,8 @@ export const _procError = err => {
 export function isAuth() {
   return !!localStorage.getItem('auth-token')
 }
+
+export function getUserType() {
+  let token = localStorage.getItem('auth-token')
+  return token ? token.substr(token.length - 1) : null
+}