+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
 - Stylus all pages & components (maxwidth = $breakpoint-xs quasar's variable)
 - Admin Default layout & Login
 - Customer Detail Page with Sign-Out feaature
+- Validator Login form[@phillip](https://github.com/hemvanh/TPTEA/pull/23)
+- Routing algorithm with new token of Admin & Customer[@phillip](https://github.com/hemvanh/TPTEA/pull/33)
+- Automation test routing [@phillip](https://github.com/hemvanh/TPTEA/pull/33)
 - Schema Admin (res loginAdmin,def type Admin,input AdminLoginInput)[@john]
 ### Removed 
 - Join now button